# polaflix/src/main/java/es/unican/mmt883/polaflix/dto/CapituloDTO.java
package es.unican.mmt883.polaflix.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CapituloDTO {
    private Long idCapitulo;
    private String titulo;
    private String descripcion;
    private int numCapitulo;
    private String enlace;
}