# polaflix/src/main/java/es/unican/mmt883/polaflix/model/Persona.java
package es.unican.mmt883.polaflix.model;

import lombok.AllArgsConstructor;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.Setter;
import lombok.NoArgsConstructor;
import jakarta.persistence.*;

@Entity
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@EqualsAndHashCode(of = "idPersona")
public class Persona {

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long idPersona;
    
    @Column(nullable = false)
    private String nombre;
    @Column(nullable = false)
    private String primerApellido;
    private String segundoApellido;
}